+++ CloudRepublic.BenchMark.sln
@@ -45,8 +45,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CloudRepublic.BenchMark.API
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Shared", "Shared", "{7572AE3C-D0EC-4843-9952-EEDD0F0DC68C}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CloudRepublic.BenchMark.Tests.Infrastructure", "tests\Shared\CloudRepublic.BenchMark.Tests.Infrastructure\CloudRepublic.BenchMark.Tests.Infrastructure.csproj", "{41137D3F-C062-446D-B79B-EB830C403855}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Web", "Web", "{CAF17107-F4CF-4D2A-8B9D-7B0D7BE4B78B}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Deployment", "Deployment", "{61317263-9BD7-4C03-A835-706C21500683}"
@@ -66,6 +64,8 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CloudRepublic.BenchMark.Orchestrator.Tests", "tests\Orchestrator\CloudRepublic.BenchMark.Orchestrator.Tests\CloudRepublic.BenchMark.Orchestrator.Tests.csproj", "{FBE92687-0D62-4E72-9248-0A3907E1DA14}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CloudRepublic.BenchMark.Application.Tests", "tests\Shared\CloudRepublic.BenchMark.Application.Tests\CloudRepublic.BenchMark.Application.Tests.csproj", "{8FE1199A-D16E-46AF-82BE-776B7A73481A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CloudRepublic.BenchMark.Test.Helpers", "tests\Shared\CloudRepublic.BenchMark.Test.Helpers\CloudRepublic.BenchMark.Test.Helpers.csproj", "{D5F0C005-7C85-4C5E-BC5D-7A657E134A15}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -105,10 +105,6 @@ Global
 		{921FC096-0A7C-470A-9384-64C80534CA75}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{921FC096-0A7C-470A-9384-64C80534CA75}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{921FC096-0A7C-470A-9384-64C80534CA75}.Release|Any CPU.Build.0 = Release|Any CPU
-		{41137D3F-C062-446D-B79B-EB830C403855}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{41137D3F-C062-446D-B79B-EB830C403855}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{41137D3F-C062-446D-B79B-EB830C403855}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{41137D3F-C062-446D-B79B-EB830C403855}.Release|Any CPU.Build.0 = Release|Any CPU
 		{AEA34EBD-B83D-4B19-87CF-48835ADF75C5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{AEA34EBD-B83D-4B19-87CF-48835ADF75C5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AEA34EBD-B83D-4B19-87CF-48835ADF75C5}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -125,6 +121,10 @@ Global
 		{8FE1199A-D16E-46AF-82BE-776B7A73481A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8FE1199A-D16E-46AF-82BE-776B7A73481A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8FE1199A-D16E-46AF-82BE-776B7A73481A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D5F0C005-7C85-4C5E-BC5D-7A657E134A15}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D5F0C005-7C85-4C5E-BC5D-7A657E134A15}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D5F0C005-7C85-4C5E-BC5D-7A657E134A15}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D5F0C005-7C85-4C5E-BC5D-7A657E134A15}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -149,7 +149,6 @@ Global
 		{CEDEECB3-F5BD-4A53-B069-530E2CAD9489} = {3D4FD3F0-1F1F-4B24-B6B7-B73C8A0385FA}
 		{921FC096-0A7C-470A-9384-64C80534CA75} = {CEDEECB3-F5BD-4A53-B069-530E2CAD9489}
 		{7572AE3C-D0EC-4843-9952-EEDD0F0DC68C} = {3D4FD3F0-1F1F-4B24-B6B7-B73C8A0385FA}
-		{41137D3F-C062-446D-B79B-EB830C403855} = {7572AE3C-D0EC-4843-9952-EEDD0F0DC68C}
 		{CAF17107-F4CF-4D2A-8B9D-7B0D7BE4B78B} = {83A6DE47-D2F3-46C9-8534-EE64C64D1F01}
 		{61317263-9BD7-4C03-A835-706C21500683} = {83A6DE47-D2F3-46C9-8534-EE64C64D1F01}
 		{AEA34EBD-B83D-4B19-87CF-48835ADF75C5} = {61317263-9BD7-4C03-A835-706C21500683}
@@ -160,6 +159,7 @@ Global
 		{5B70698A-1EB4-4A41-A1E8-DA6A87EEEB62} = {3D4FD3F0-1F1F-4B24-B6B7-B73C8A0385FA}
 		{FBE92687-0D62-4E72-9248-0A3907E1DA14} = {5B70698A-1EB4-4A41-A1E8-DA6A87EEEB62}
 		{8FE1199A-D16E-46AF-82BE-776B7A73481A} = {7572AE3C-D0EC-4843-9952-EEDD0F0DC68C}
+		{D5F0C005-7C85-4C5E-BC5D-7A657E134A15} = {7572AE3C-D0EC-4843-9952-EEDD0F0DC68C}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {0580A24C-4FAC-46B1-8BAA-78D167E4D0AF}
