+++ tests/API/CloudRepublic.BenchMark.API.Tests/CloudRepublic.BenchMark.API.Tests.csproj
@@ -7,6 +7,7 @@
     </PropertyGroup>
 
     <ItemGroup>
+        <PackageReference Include="Moq" Version="4.11.0" />
         <PackageReference Include="xunit" Version="2.4.0" />
         <PackageReference Include="xunit.runner.visualstudio" Version="2.4.0" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.9.0" />
@@ -14,7 +15,7 @@
 
     <ItemGroup>
       <ProjectReference Include="..\..\..\src\API\CloudRepublic.BenchMark.API\CloudRepublic.BenchMark.API.csproj" />
-      <ProjectReference Include="..\..\Shared\CloudRepublic.BenchMark.Tests.Infrastructure\CloudRepublic.BenchMark.Tests.Infrastructure.csproj" />
+      <ProjectReference Include="..\..\Shared\CloudRepublic.BenchMark.Test.Helpers\CloudRepublic.BenchMark.Test.Helpers.csproj" />
     </ItemGroup>
 
 </Project>
