+++ tests/Shared/CloudRepublic.BenchMark.Test.Helpers/MockILoggerFactory.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Logging.Abstractions;
+
+namespace CloudRepublic.BenchMark.Test.Helpers
+{
+    public class MockILoggerFactory
+    {
+        public static ILogger CreateLogger()
+        {
+            return NullLoggerFactory.Instance.CreateLogger("Null Logger");
+        }
+    }
+}
