--- tests/Shared/CloudRepublic.BenchMark.Tests.Infrastructure/CloudRepublic.BenchMark.Tests.Infrastructure.csproj
@@ -1,20 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>netstandard2.0</TargetFramework>
-        <RootNamespace>CloudRepublic.BenchMark.Tests.Infrastructure</RootNamespace>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <PackageReference Include="Microsoft.AspNetCore.Http" Version="2.2.2" />
-      <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.4" />
-      <PackageReference Include="Microsoft.Extensions.Logging" Version="3.1.4" />
-      <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="3.1.4" />
-      <PackageReference Include="Moq" Version="4.11.0" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\..\..\src\Shared\CloudRepublic.BenchMark.Application\CloudRepublic.BenchMark.Application.csproj" />
-    </ItemGroup>
-
-</Project>
